Add Dockerfile, .gitignore, etc.
Added the various things that are required. I've actually overhauled what I used to be doing, and an older version of the current Dockerfile can be found in the dockerfiles directory. It won't actually work though because it requires some other resources, but honestly that thing is such a mess, that I'll probably end up deleting it at some point.

The Dockerfile at the project root is much more sane, and has a decent (I hope) amount of commenting too. The other one also has a decent amount of commenting, but it also has a decent amount of useless, obsolete stuff.