Try to upgrade python3.5 to python3.6, yet build fails
Build doesn't seem to be improved by making python3.6 the main version of python running on the Docker image,
so I think that it would probably be best if we just went back to using melodic...and somehow built the
opencv3 package for melodic. I'm not sure what would happen if we don't have the opencv3 package, but it has to
be better than dealing with Ubuntu 16.04 being dependent on python3.5.

Resolve #5, since I don't believe I want to deal with different python
versions very much aha.