Revert to melodic, attempt opencv3 build
Decided to stop dealing with the nonsense involved in trying to build using the ros:kinetic image and just
store away the associated dockerfile in the dockerfiles directory, and moved back out the Dockerfile-melodic
version to just use as our regular Dockerfile. This is related to
issue #5.

I've added some commands that are supposed to build opencv3, but it seems to take up a lot of space/memory
which has exhausted all of the memory/storage available to the docker container. Is there some kind of
upper limit or something that I can increase to make sure that the docker container doesn't run out of
memory so easily? Documented in issue #6.