Create different Dockerfiles for w/wo opencv3 build
Split up the Dockerfiles so that we have two options to choose from, I suppose.
This is related to #4.

Additionally, played around a little bit with how much memory was allotted to
Docker, and it seems like 4GiB of memory both prevents opencv3 build from crashing
(maybe) and also seems like it resolves the problem of having to build cartographer_ros
twice, because it builds it much faster, #7, although I'd have to test it again
to be sure that that was the cause.