Add configuration files and launch files
The files my-bash_profile and my-bashrc just provide basic good things for the docker container,
like sourcing all of the ROS-related setup, which makes testing the launch files I've been trying to test out easier.

Also trying to test out the X stuff, which is what the file dockerfiles/Dockerfile-gui-firefox-demo is for.