Add rviz-diff directory, modify READMEs
Added the rviz-diff directory that automatically finds similar-sounding
files from a ros1 version and tries to find the ros2 version of those files.
This gets around the fact that git's own diff tool tends to believe that
they're simply deleted, instead of being moved somewhere else and then altered.

I added a README to the rviz-diff directory that explains how to run it
and where to find the output.

I also modified the main README a little bit, because there was a sentence fragment.