Finally figure out X forwarding
I've been trying to run the rviz stuff, but that means that I have to X forwarding
because rviz does GUI stuff. I ended up having to scroll deep into the comments
of one of the tutorials to find a fix for this.

But now rviz doesn't want to actually do anything because it needs a lot more
imaging power, so it needs to use my graphics card, which my laptop doesn't
want to do.