Add launch file, add pkg spec json, modify dockerfiles
Added the launch file for convenience, and modified some Dockerfiles so that it would create a ROS2 package workspace as-needed for us.

Also added the package spec json, and I'll be adding some of the code transformations soon -- the pkg spec json needs to be organized more into actual include directives vs snippets of code.