Figure out the ros1 version
Made a specific dockerfile for running the ros1 version of the test from the pdf.
This involves a somewhat more involved build, and it probably is not best practices
as regarded by Docker, but it's basically faster.

I've also added some of the files necessary for launching and running the program.
It's been modified so that it only prints out 10 times, and it exits.