Update README, make .dockerignore more useful
Updated README to reflect the various different dockerfiles that are in this repo before I start to forget what some of these are for.
I only described the ones that actually work though.

I have the .dockerignore finally ignore emacs temp files, which is really great.

I also noted the inconsistency between how the Dockerfiles are named and what the names end up being when they're built, in the README,
which closes #11.