Fix typos, verify the test1_ros2 case works
There were a couple of typos in Dockerfile-bin-melodic that I had to clear up.

Also added development dependencies into the Dockerfile-x-dev-melodic, which made the
build longer, but I was able to build with colcon build --symlink-install and then
after sourcing the install/setup.bash script, I could launch it with no problems! This
resolves #10, and also resolves #9 since both of the test cases from the pdf now work.