Figure out how to get ros nodes to exit, launch to stop
So I could call exit on them, but the process wouldn't actually stop on its own unless I did
control-C, and apparently you can make the node required (if it's the xml launch files of ros1)
or you can just register an event handler for shutdown (I guess the default is that there are
some nodes, and nodes can just go in and out so it doesn't matter if nodes go away). Anyway, this
fixes #13.