Add more to pkg spec, add sketch and code spec
I created a sketch out of the listener file that I've also annotated with some reasoning
that whatever program I'm going to code eventually will be doing, hopefully. Working on #16.

Also added the specification for the code changes from ROS1 to ROS2, and
I grouped them by the type of changes they'd be, whether we're replacing
a function call or a type annotation. So I'm working on #15 as well.