Add build file, create json parser
Using eclipse to write the java code that parses the listener-search-space.json
file, so there's some things in there for that. I made the root of the project
the folder for the workspace in eclipse...which leaves something to be desired.

Added ros2sy folder, which contains all of the code for ros2 synthesis, or at
least it will eventually. The json parser, that will digest json files with
the descriptions of API's, is functional, although it only works on one file
at the moment. Also added some classes that allow one to describe methods
easily, and this is what essentially will hold the results from the json file.