Create petri net from parsed methods
Added a class that takes a list of methods and makes a petri net from it.
It still doesn't have all of the tokens and other things, such as the
transitions that 'copy' the variable so that you can use it again, and
other things like that. However, it does take into account whether certain
arguments to a function have defaults, and makes alternate branches from them.