Add more coverage of C++ types in petri net
The petri net was previously not taking into account that instances of
classes needed to be present in order to be able to use certain methods.
Methods were changed to contain a field for the class of the instance
they need, if any. This fixes #19.

Shared pointers are a common type in ROS code that can also be used to
access such methods. In order to denote this, and to provide some
Petri Net concept of the pointer notation, since that makes a difference
in the code, transitions were added for the pointer notation in a shared
pointer, which is relevant to #18, since this is along the lines of
adding the additional transitions to the petri net.

Additionally, more documentation was added, and skeletons of javadoc
comments were at least added to most of the code files.