Assign max tokens, update build file
Build file needed to have each of the libraries listed, with SAT4j listed first,
since otherwise Java will not know where SAT4j is, essentially.

Also added in assigning max tokens, which was absent in the version of the
library that I had previously, but now that it's added, this actually compiles
annd runs.