Reorganized package, clean up signature code
Decided that the actual code-outputting part should be contained in the ros2sy.code package,
and the classes that represent method signatures are now located in the ros2sy.sig package.

Cleaned up a lot of the code in MethodsToPetriNet, since that's the biggest file in the package
thus far. I also made it so that it wasn't entirely just static methods, because it turned out
that having some sense of state would be good for turning the PetriNet into code.