Return solns from synthesis, start C++ code gen
Modified the Synthesis class so that it returned an ArrayList of
ArrayLists of Strings, so that we can analyze them and use them
to create our little C++ programs.

To start working on C++ code generation (#22), we first need a program
that can generate the code with the holes. This commit adds that
functionality. For example, we were able to generate this funny little
bit of code:

[java] rclcpp::Context::init(# 0, # 1);
[java] bool the_freshest_id_0 = rclcpp::shutdown();
[java] bool the_freshest_id_1 = rclcpp::shutdown();

It puts in the holes for the arguments, and generates fresh variables
for the results of non-void functions.