Add hole-filling combinatorial search
This is most of the second stage of issue #22, for the C++ code
generation. I'm currently testing it on a sequence of API calls that it
can't actually generate itself, or at least it doesn't seem practical
for me to sit there and wait for it to do so at the moment, and gives it
a sequence of inputs that I know to be the "correct" inputs, from the
actual ROS2 version.

The hole-filler that generates the code has two stages. First, the holes
that are known to only have one reasonable input are filled with that
input. Second, the remaining holes with more than one input are filled
using a recursive function, and all possible hole-fills are produced.

The only part that is left is how to indicate when it can use the
results of previous API calls.