Put apis into blocks, use results in hole-filling
The hole-filler can now use previous results in filling holes. This
works on #22.

Since the whole task of synthesis is going to be split up into chunks by
blocks, I've created three sets of methods, one of which must appear in
each chunk. These three sets are contained within an ArrayList, so we
can iterate through this ArrayList to synthesize the code.