Fix small issue with sort, test out blocks
The sort was never returning 0, which meant that it was always switching
sequences of APIs even if they were equivalent in cost and length, so I
just a clause for the case where they were equal.

Blocks work great for init and spin, but not so much for the
node/create_subscription part, and we're not even looking at all of the
public APIs for the rclcpp::Node or rclcpp::Subscription namespaces.
Likely, we need to be able to exclude groups of APIs, as much as
including at least one of other groups of APIs. This works further on
issue #22.