Create ROS2 doc scraping script
This works on #27.

This script needs to be changed though so that it scrapes from the more
verbose sections of the documentation, so that we can have some added
information, such as the template types -- aka information about
parametric polymorphism. We probably won't get to that during this
summer, but I think it would be just very nice to have for the future.
Plus, parsing from the more verbose part should also make the actual
parsing script cleaner.