Resolve #27: scraping functions from ROS2 docs
This scraping script is fairly extensible so in the future, if the
search space is to be expanded, it should be fairly easy to do so,
hopefully. Part of the benefit of doing this is that it's easy to update
the search space to the latest version of ROS2: as long as they don't do
something like reorganize the website (unlikely since it's generated by
doxygen in the first place), we should be able to get the latest version
always by just re-running the scraping script.

More work has to be done though to integrate the jsons generated from
scraping the ROS2 docs into the actual synthesis code.