Move main class to Synthesis, and tag methods
I've added a somewhat flexible/extensible way of adding tags to methods,
which the scraper generates. I've also made it so that we can easily get
the intersections between these sets of tagged methods, which will be
cool.

The main method in Synthesis doesn't run the actual synthesis, yet, but
this mainly fixes #29, the rudimentary annotations for the analysis.

Resolves #30, since I figured out why the permalink wasn't correct for
overloaded functions, besides the first one: it never gave an updated
link for each succesive overload, if you will.

This also fixes #31, the fact that rclcpp::Node::make_shared isn't even
in the documentation, by just adding it manually.