Fix #32: class method petri net bug
Finally got around to fixing #32, which means that now many of the class
methods don't think that they need TWO instances of whatever object in
order to be used, and now they only need one.

After running the synthesizer, it becomes evident that our answer is
actually in the top 10 every single time: it's the first one for the
init block, the first one for the node/subscription block, and the sixth
one for the spin block.