Reintegrate hole-filling with synthesis
This fixes #37 by adding back in hole-filling to the things that happen
in the main function, essentially.

This also resolves #34, since the Type class now recognizes array types.

Additionally, the input types and other information that was previously
hard-coded in have been moved into json files that are in the inputs
directory.