Fix #35: more direct 'type casts'
Now there are transitions added that directly bring one type to an
equivalent type, instead of needing two or more jumps in between.

However, this has seemed to make dot hate me, since it no longer renders
all of the nodes in any sort of detail, and trying to render a png seems
like it will take forever.

The good news is, though, that now we can just limit the search space to
3 transitions.