Resolve #40: re-add sketch hole filling
Fixed #40 by creating a new class called a SketchFiller that generates
the C++ code and uses it to fill in the designated sketch. However, it
does not replace the includes. This works on #22, for C++ code gen.

Also re-did the formatting of holes for the listener sketch, so now I'm
using ?#? as the hole tag.

Realized that there was a bunch of junk in the ros1-to-ros2-code-spec
json file, which I have now removed.