Resolve #25: automagically add include directives
Every single function has one of these include directives, or at least
it does in the json, so we simply need to retrieve the proper include
directive and put it in a reasonable place.

This code also corrects incorrect include directives, and doesn't add
duplicates. However, we don't necessarily add the minimum number of
includes, since we only need to make sure that everything will actually
compile.

This finally uses my ros1-to-ros2-code-spec.json and
ros1-to-ros2-pack-spec.json files to convert some ros1 includes and
types to ros2 types.

This makes some leaps with working on #22, which is all of codegen.

However, the code still doesn't compile since we don't get rid of the
"fake" types such as SubscriptionT, and some methods require templating
as well, which I discovered when I tried to run a build. This has
resulted in the new issue #42.