Resolve #41: create SearchSpace class
The SearchSpace class acts as a convenient interface with the sets of
names of methods, which are organized by tags.

This allows us to easily get intersections between them, using helper
methods.

Additionally, the SearchSpace class can also provide a way to organize
the blocks of code, which is another part that added complexity to the
original program.

This has made the main function a lot more readable and concise.