Resolve #43: implement logging
Now it helpfully creates logs and stores them in a log file while only
printing out the very necessary stuff, so that we can keep track of
what's happening when.

Also, I decided to actually use debug with javac so that we can have
line numbers and such while debugging this thing, which is good.