Get 6/7 blocks in talker working
I think that the remaining block that doesn't work is simply too big.

It turns out that the reason why the last block wasn't working is that I
forgot to account for overlap between the set of things that we do want
to occur and the set of things that we don't want to occur, so now I
make sure that they are mutually exclusive so that we don't have a
contradiction, basically, in our constraint solver.

I also made some headway for how things are formatted, so class methods
have the `rclcpp::Node` or whatever preceding part chopped off, which
looks nice.