Fix parsing bug with return types
Return types were being messed up in functions that had templates,
because the space right after the comma would be confused for the space
in between the return type and the name of the function.

I'm trying to add more support for template types, since that seems to
be a problem with trying to get the publisher to actually publish the
message that we generate, so this goes to working on #51.