Work more on templates, fix #53
Added parsing of templates, which resolves #52. Additionally adds some
classes for organizing templates as objects, and actually uses some
object oriented-ness, finally, so that we can define new "types" of
templates, for example MessageT, which wants a type that has "::msgs::"
in it, so working on #51.

Also, I've been running into this weird bug (#53) which has prevented
the synthesizer from synthesizing any code that is about the Publisher,
which is very odd. It turned out that because I was getting an instance
of the set of places from the petri net before ADDing more places to the
petri net, only after which I would finally give all of the places a max
token, many of the newly added places would have a max token of 0.
Therefore, in the loop that adds the max token, I just replaced the
reference to the set of places with a call to the method to get the set
of all places, so that it will always be up to date. Thus this closes