Add more template type/type availability fxns
Added some more functions for figuring out which template parameters are
required for a given Method, which is related to #51.

Also parsed additional type information from the docs using the scraper,
which is nice because it gives us the types that become available when
we include a certain ROS file. It doesn't recursively check all of the
included headers, though, but that should be fine for our purpsoses
right now. This closes #56.