Find available MessageT types
Synthesizer now finds the available types that can fit a MessageT, which
I've defined as all types matching the regex /::msgs::/, since all ROS2
message types have to be nested within that messages namespace or
something. If there's only one such type messageType, then when we
generate the petri net, we replace all occurrences of MessageT with
messageType. This provides the foundation for working on #57, but
unfortunately I don't think it's very extensible without adding
functions as first-class values since we have to check if a given type t
is similar to a MessageT, PublisherT, etc. which all take different
predicates, I'd imagine. I might be able to just make some kind of
dictionary but this would largely be hard-coding it.

I also switched to using IntelliJ Idea because Eclipse was compiling
immensely slowly. IntelliJ Idea is compiling faster but is doing some
strange things, which may be due to either switching IDEs (which could
change the compiler it's using) or due to changing the code. It's only
generating about 2/8 of the code blocks correctly.

Since I switched to IntelliJ Idea I had to move the resources folder so
that it wasn't nested inside the src folder, because it won't allow them
to be nested.

The ant build.xml file was also refactored to use the ant.project.name
property instead of repeatedly writing out "ros2sy" everywhere, just in
case we change the name of the project.