Generalize parameter replacement for #57
Parameter replacement has been generalized by creating a json file --
inputs/param-map.json -- that contains the parameters that you wish to
replace, and the regex for recognizing types that fit that parameter.
This could be relaxed in order to have arrays of regexes, so that it
only has to fit one of the regexes in order to match, and it's fairly
easy to add more parametric type variables to the json file in order to
extend this. This is at least a less evil option than having an if
statement with a variable number of else-if blocks following it, and
it's all in one place.

For right now, parameters are only replaced if there is only one option,
since that's the easiest case.

The map of regexes is available as a static member of TemplateParameter.
An additional method was created to replace parametric type vars in Type
objects.

Additionally, there was a bug that was causing very little code to get
generated, and most of the code that was generated was either trivial or
not very correct. Apparently, this was due to the fact that I was
re-creating Method objects incorrectly after replacing all of the
instances of parametric type vars in them. Their args, specifically,
weren't being labeled as having default values, so every method only
corresponded to one transition in the petri net, unless there were
overloaded functions. This has been fixed in this commit by changing how
args are represented as strings, which fixes #58.