Remove IntelliJ Idea workspace config file
It was recording silly things that don't need to be put on the repo itself. I didn't add it to the gitignore file because not everyone using this project is using IntelliJ Idea, and this was more specific to the machine than anything. If more people decide to use IntelliJ Idea with this project, then we can add it to the gitignore.