Finally get publish to synthesize
For some reason publishing would not synthesize, which I couldn't
understand for the life of me. It came down to mismatched angle
brackets, which I tracked down to some dumb code in the Type class. It
was improperly taking a substring between two angle brackets, and would
just go to the end of the string instead of to the last angle bracket.
This closes #59.

However, this doesn't mean that it actually fully synthesizes the code,
but at least the # 9.publish(# 10) is in there for me to replace.