Get rclcpp::Publisher::publish to synthesize
Finally got it to synthesize. I had it replace all of the type params in
most of the return types, and also the fromClass field of Method
objects.

This means that most of #61 is done now, I just need to make sure that
it actually uses the template parameters and puts them into the code.