Fix #47, get the synthesizer to create talker
The talker finally synthesizes. This was a combination of things, such
as making sure that the right types were put in the right places, and
also putting in reasonable types for the holes created by parametric
types. Note that I have not tried compiling though yet.

I also inadvertently fixed #62, since I finally got the
rclcpp::GenericRate to go away and just become rclcpp::WallRate, an
alias of GenericRate really, by writing the wallrate.py script which
takes the GenericRate methods and makes basically a copy of them that is
their own.

This script also adds the tags that WallRate should have to the tags
json file.