Rename file for listener test, fix listener
Renamed the dockerfile for running the listener test so that it's
consistent with the dockerfile for running the talker one.

Fixed one of the bugs with the listener, which was that the type
params in std::shared_ptr<rclcpp::Subscription<CallbackMessageT, Alloc>>
weren't being substituted properly, which has now only left us with the
bug of the return types that are inserted being somewhat incorrect,
which is documented as #65.