Fix #45, fix #65 -- make listener compile
The listener now compiles since we're using the auto keyword to our
advantage (#45) and this has silenced a good number of the listener
compiler error messages (#65).

Also found and fixed another bug where we erroneously gave
`rmw_qos_profile_system_default` the type of `rclcpp::QoS`, and gave it
its actual type of `rmw_qos_profile_t`, which made it so that it used
the proper overloaded version of `rclcpp::Node::create_subscription`.
This in turn triggered a bug that was previously unknown to me because
it was a dynamic improper regex error, and previously, that code had
never been run since no optional arguments were being used until that
point,  and an accidental dead code-triggered error, since I wasn't
re-assigning the result of a java.lang.String.replaceAll back to the
original string, since replaceAll doesn't actually act on the string
itself.

So now listener compiles, but since it erroneously uses
rclcpp::spin_some instead of rclcpp::spin, the listener node exits early
and so the whole thing is basically wrong. This will be fixed by #66.